+++ spec/foundation_rails_helper/configuration_spec.rb
@@ -18,18 +18,42 @@ describe FoundationRailsHelper do
       end
     end
 
+    describe "#ignored_flash_keys" do
+      it "defaults to empty" do
+        config = FoundationRailsHelper::Configuration.new
+        expect(config.ignored_flash_keys).to eq([])
+      end
+    end
+
+    describe "#ignored_flash_keys=" do
+      it "can set the value" do
+        config = FoundationRailsHelper::Configuration.new
+        config.ignored_flash_keys = [:foo]
+        expect(config.ignored_flash_keys).to eq([:foo])
+      end
+    end
+
     describe ".reset" do
-      before :each do
+      it "resets the configured button class" do
         FoundationRailsHelper.configure do |config|
           config.button_class = 'new-class'
         end
-      end
 
-      it "resets the configuration" do
         FoundationRailsHelper.reset
 
         config = FoundationRailsHelper.configuration
         expect(config.button_class).to eq('small radius success button')
+      end
+
+      it "resets the configured ignored flash keys" do
+        FoundationRailsHelper.configure do |config|
+          config.ignored_flash_keys = [:new_key]
+        end
+
+        FoundationRailsHelper.reset
+
+        config = FoundationRailsHelper.configuration
+        expect(config.ignored_flash_keys).to eq([])
       end
     end
   end